exoplanets modal refactored
