+++ src/components/exoplanets/inspect-exoplanets.js
@@ -1,7 +1,7 @@
 "use client";
 import { fetchBatchedExoplanets } from "@/services/fetch-datasets";
 import { useEffect, useState } from "react";
-import ExoplanetModal from "./exoplanet-modal";
+import Modal from "../modal";
 
 export const InspectExoplanets = () => {
   const page_size = 15;
@@ -41,15 +41,6 @@ export const InspectExoplanets = () => {
   const goPrev = () => {
     setPage((page) => Math.max(0, page - 1));
   };
-
-  // close modal on Escape
-  useEffect(() => {
-    const onKeyDown = (event) => {
-      if (event.key === "Escape") setSelectedPlanet(null);
-    };
-    window.addEventListener("keydown", onKeyDown);
-    return () => window.removeEventListener("keydown", onKeyDown);
-  }, []);
 
   return (
     <>
@@ -113,8 +104,83 @@ export const InspectExoplanets = () => {
       </div>
   </section>
 
-  {/* Modal for selected planet */}
-  {selectedPlanet && <ExoplanetModal selectedPlanet={selectedPlanet} onClose={() => setSelectedPlanet(null)} />}
+  {/* Modal for selected planet (inlined) */}
+  {selectedPlanet && (
+    <Modal
+      isOpen={!!selectedPlanet}
+      onClose={() => setSelectedPlanet(null)}
+      title={
+        (function formatValue(v) {
+          if (v === null || v === undefined || v === "") return "—";
+          return String(v);
+        })(selectedPlanet.planet_name || selectedPlanet.name || "Unnamed")
+      }
+    >
+      {(() => {
+        const formatValue = (value) => {
+          if (value === null || value === undefined || value === "") return "—";
+          return String(value);
+        };
+
+        const formatControversial = (value) => {
+          if (value === 0) return "False, the discovery of this exoplanet is not disputed by the Space Community.";
+          if (value === 1) return "True, the discovery of this exoplanet is disputed by the Space Community.";
+          if (value === null || value === undefined || value === "") return "-";
+        };
+
+        return (
+          <div className="mt-0 bg-gray-50 dark:bg-gray-900 p-3 rounded text-sm">
+            <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
+              <div>
+                <div className="text-xs text-gray-500">Discovery year</div>
+                <div className="text-sm text-gray-800 dark:text-gray-200">{formatValue(selectedPlanet.discovery_year)}</div>
+              </div>
+
+              <div>
+                <div className="text-xs text-gray-500">Discovery method</div>
+                <div className="text-sm text-gray-800 dark:text-gray-200">{formatValue(selectedPlanet.discovery_method)}</div>
+              </div>
+
+              <div>
+                <div className="text-xs text-gray-500">Discovery facility</div>
+                <div className="text-sm text-gray-800 dark:text-gray-200">{formatValue(selectedPlanet.discovery_facility)}</div>
+              </div>
+
+              <div>
+                <div className="text-xs text-gray-500">Instrument</div>
+                <div className="text-sm text-gray-800 dark:text-gray-200">{formatValue(selectedPlanet.discovery_instrument)}</div>
+              </div>
+
+              <div>
+                <div className="text-xs text-gray-500">Controversial?</div>
+                <div className="text-sm text-gray-800 dark:text-gray-200">{formatControversial(selectedPlanet.controversial_flag)}</div>
+              </div>
+
+              <div>
+                <div className="text-xs text-gray-500">Orbital period (days)</div>
+                <div className="text-sm text-gray-800 dark:text-gray-200">{formatValue(selectedPlanet.orbital_period_days)}</div>
+              </div>
+
+              <div>
+                <div className="text-xs text-gray-500">Semi-major axis (AU)</div>
+                <div className="text-sm text-gray-800 dark:text-gray-200">{formatValue(selectedPlanet.orbital_semi_major_axis_in_au)}</div>
+              </div>
+
+              <div>
+                <div className="text-xs text-gray-500">Radius (Earth radii)</div>
+                <div className="text-sm text-gray-800 dark:text-gray-200">{formatValue(selectedPlanet.radius_earth_radii)}</div>
+              </div>
+
+              <div>
+                <div className="text-xs text-gray-500">Star radius (Solar radii)</div>
+                <div className="text-sm text-gray-800 dark:text-gray-200">{formatValue(selectedPlanet.star_radius_solar_radii)}</div>
+              </div>
+            </div>
+          </div>
+        );
+      })()}
+    </Modal>
+  )}
     </>
   );
 };
